fix(API): Update api controllers
