+++ src/routes/api/posts/getPosts.ts
@@ -1,0 +1,31 @@
+import type { Post } from '$lib/types';
+
+type Filters = Record<'limit', string>;
+
+export async function getPosts(filters: Filters = {} as Filters) {
+	let posts: Post[] = [];
+	const limit = parseInt(filters.limit, 10);
+
+	const paths = import.meta.glob('/src/posts/*.md', { eager: true });
+
+	for (const path in paths) {
+		const file = paths[path];
+		const slug = path.split('/').at(-1)?.replace('.md', '');
+
+		if (file && typeof file === 'object' && 'metadata' in file && slug) {
+			const metadata = file.metadata as Omit<Post, 'slug'>;
+			const post = { ...metadata, slug } satisfies Post;
+			post.published && posts.push(post);
+		}
+	}
+
+	posts = posts.sort(
+		(first, second) => new Date(second.date).getTime() - new Date(first.date).getTime()
+	);
+
+	if (!isNaN(limit)) {
+		posts = posts.filter((_, idx) => limit > idx);
+	}
+
+	return posts;
+}
