+++ src/routes/api/workbench/+server.ts
@@ -1,41 +1,9 @@
 import { json } from '@sveltejs/kit';
-import type { WorkbenchMetadata } from '$lib/types';
-import { getFilterFromURLBy } from '$lib/utils';
+import { getWorkbench } from './getWorkbench';
 
-type Filters = Record<'limit', string>;
-
-async function getWorkbench(filters: Filters) {
-	let posts: WorkbenchMetadata[] = [];
-	const limit = parseInt(filters.limit, 10);
-
-	const paths = import.meta.glob('/src/workbench/*.md', { eager: true });
-
-	for (const path in paths) {
-		const file = paths[path];
-		const slug = path.split('/').at(-1)?.replace('.md', '');
-
-		if (file && typeof file === 'object' && 'metadata' in file && slug) {
-			const metadata = file.metadata as Omit<WorkbenchMetadata, 'slug'>;
-			const post = { ...metadata, slug } satisfies WorkbenchMetadata;
-			post.published && posts.push(post);
-		}
-	}
-
-	posts = posts.sort(
-		(first, second) => new Date(second.date).getTime() - new Date(first.date).getTime()
-	);
-
-	if (!isNaN(limit)) {
-		posts = posts.filter((_, idx) => limit > idx);
-	}
-
-	return posts;
-}
-
-export async function GET({ request }) {
+export async function GET() {
 	try {
-		const limit = getFilterFromURLBy(request.url, 'limit');
-		const workbench = await getWorkbench({ limit });
+		const workbench = await getWorkbench();
 
 		return json(workbench);
 	} catch (error) {
